API: separation of fields and mappers (#86)
This is a fairly large change to the code that separates the
functionality of fields and mappers (e.g. for HEALPix maps).

Randomisable fields have been removed, and fields are always normalised.

Factoring out the mapper functionality into its own class is not
complete in some parts of the code, such as e.g. the `twopoint` module.

Closes: #85