API(fields): turn spin weight into a class property (#93)
Turn the spin weight of `Field` subclasses from an `__init__()` keyword
into a metaclass keyword. This removes the need for a trivial
`__init__()` override in subclasses.

By default, the spin weight is undefined. If an undefined spin weight is
obtained via the default `.spin` property, a `ValueError` is raised.

If necessary, `Field` subclasses can override the default `.spin`
property to implement dynamic, instance-based spin weights.

Closes: #92
Closes: #51