ENH: command line interface (#72)
Implementation of the Heracles command line interface (CLI). The CLI is
meant to make it quick and easy to perform a "standard" analysis of
catalogues, that is: obtaining angular power spectra, mixing matrices,
and eventually jackknife covariances.

The CLI uses a configuration file (default file name `heracles.cfg`), of
which an example is provided in the `examples/` folder.

The analysis itself uses a series of commands that build up e.g. angular
power spectra:

    heracles alms myalms.fits
    heracles spectra myspectra.fits myalms.fits
    heracles mixmats mymixmats.fits myalms.fits

The following commands are currently implemented:

* `maps` -- create HEALPix maps from catalogues
* `alms` -- compute spherical harmonic coefficients (alms); either from
catalogues or pre-computed maps
* `spectra` -- compute angular power spectra from alms
* `mixmats` -- compute mixing matrices

There is help available for every command (`heracles <cmd> -h`).

Closes: #56