API: simplify Mapper interface (#120)
Simplifies the `Mapper` interface so that it no longer relies on
metadata in the outputs. All in all, a lot of code is now significantly
simpler and easier to follow.

Each `Field` now knows about its associated mapper, similar to how it
was with `nside` before the `Mapper` interface was introduced.

Implementations of the `Mapper` interface are now structural instead of
nominal subtypes. The `Mapper` class itself is now a `Protocol`.

The mapping functions `map_catalogs()` and `transform_maps()` now
receive only the list of fields and their data. The `Healpix` mapper
itself keeps track of `lmax` and whether or not to deconvolve the pixel
window function. This allows to make it a per-field property.

In the two-point functions, the debiasing step now special-cases HEALPix
maps explicitly, without relying on the reconstruction of the mapper
that created each alm.

Closes: #119